requirements.txt
Django==5.2.9
Pillow>=10.0.0

